I have to load more the CPU ? (and python doesn't work)

+++ include/PPN-microbench/load_test.hpp
@@ -12,7 +12,7 @@ class LoadTest : public Microbench {
     private:
         int nbMeasures;
         int nbCores;
-        std::vector<int> measures;
+        std::vector<double> measures;
         
     public:
         LoadTest(int nbMeasures);
